requirements.txt
#python-telegram-bot[webhooks]==21.9
#gspread
#oauth2client

python-telegram-bot==20.3
gspread==5.11.0
oauth2client==4.1.3